It compiles again :D, now to add systems to plugins and the new components to the bundles
