+++ Cargo.toml
@@ -13,8 +13,9 @@ repository = "https://github.com/PROMETHIA-27/bevy_mod_wanderlust"
 version = "0.4.0"
 
 [features]
-default = []
+default = ["rapier"]
 debug_lines = []
+rapier = ["bevy_rapier3d"]
 
 [dependencies]
 bevy = { version = "0.11", default-features = false, features = [
@@ -24,7 +25,7 @@ bevy = { version = "0.11", default-features = false, features = [
 bevy_rapier3d = { version = "0.22", default-features = false, features = [
     "async-collider",
     "dim3",
-] }
+], optional = true}
 
 [dev-dependencies]
 bevy = "0.11"
