+++ src/plugins.rs
@@ -1,4 +1,4 @@
-use crate::{components::*, systems::*};
+use crate::controller::ControllerInput;
 use bevy::prelude::*;
 
 /// The [character controller](CharacterController) plugin. Necessary to have the character controller
@@ -7,11 +7,9 @@ pub struct WanderlustPlugin;
 
 impl Plugin for WanderlustPlugin {
     fn build(&self, app: &mut App) {
-        app.register_type::<ControllerState>()
-            .register_type::<ControllerSettings>()
-            .register_type::<ControllerInput>()
-            .register_type::<Option<Vec3>>()
-            .add_systems(Startup, setup_physics_context)
-            .add_systems(Update, movement);
+        app.register_type::<ControllerInput>()
+            .register_type::<Option<Vec3>>();
+        //.add_systems(Startup, setup_physics_context)
+        //.add_systems(Update, movement);
     }
 }
